pyproject.toml
[project]
name = "eeg-wm-jepa"
version = "0.1.0"
description = "Topology-Invariant EEG Foundation Model with Spatiotemporal JEPA and SIGReg"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "numpy",
    "scipy",
    "mne",
    "einops",
    "pyyaml",
]
